requirements.txt
requests
python-telegram-bot==13.17
python-dotenv